+++ nsinit/stats.go
@@ -1,4 +1,4 @@
-package main
+package nsinit
 
 import (
 	"encoding/json"
